+++ src/merc.h
@@ -2372,6 +2372,7 @@ void    quest_handle_logout ( CHAR_DATA *ch );
 /* act_move.c */
 void    move_char       ( CHAR_DATA *ch, int door, bool follow );
 ROOM_INDEX_DATA *recall_room ( CHAR_DATA *ch );
+void    recall_char_to_temple ( CHAR_DATA *ch );
 bool    room_allows_recall_point ( ROOM_INDEX_DATA *room );
 void    recheck_sneak   ( CHAR_DATA *ch );
 RID     *get_random_room( CHAR_DATA *ch );